make sure to keep archive file after it's extracted during the project config phase.

+++ src/main/java/com/dryadandnaiad/sethlans/utils/SethlansFileUtils.java
@@ -121,7 +121,7 @@ public class SethlansFileUtils {
      * @param extractLocation Location of filename
      * @return
      */
-    public static boolean archiveExtract(String toExtract, File extractLocation) {
+    public static boolean archiveExtract(String toExtract, File extractLocation, boolean deleteArchive) {
         File archive = new File(extractLocation + File.separator + toExtract);
         try {
             if (archive.toString().contains("txz")) {
@@ -129,7 +129,9 @@ public class SethlansFileUtils {
                 LOG.debug("Extracting " + archive + " to " + extractLocation);
                 Archiver archiver = ArchiverFactory.createArchiver(ArchiveFormat.TAR, CompressionType.XZ);
                 archiver.extract(archive, extractLocation);
-                archive.delete();
+                if (deleteArchive) {
+                    archive.delete();
+                }
                 return true;
             }
             if (archive.toString().contains("tar.gz")) {
@@ -137,7 +139,9 @@ public class SethlansFileUtils {
                 LOG.debug("Extracting " + archive + " to " + extractLocation);
                 Archiver archiver = ArchiverFactory.createArchiver(ArchiveFormat.TAR, CompressionType.GZIP);
                 archiver.extract(archive, extractLocation);
-                archive.delete();
+                if (deleteArchive) {
+                    archive.delete();
+                }
                 return true;
             }
             if (archive.toString().contains("tar.bz2")) {
@@ -145,14 +149,19 @@ public class SethlansFileUtils {
                 LOG.debug("Extracting " + archive + " to " + extractLocation);
                 Archiver archiver = ArchiverFactory.createArchiver(ArchiveFormat.TAR, CompressionType.BZIP2);
                 archiver.extract(archive, extractLocation);
-                archive.delete();
+                if (deleteArchive) {
+                    archive.delete();
+                }
                 return true;
             } else {
                 extractLocation.mkdirs();
                 ZipFile archiver = new ZipFile(archive);
                 LOG.debug("Extracting " + archive + " to " + extractLocation);
                 archiver.extractAll(extractLocation.toString());
-                archive.delete();
+                if (deleteArchive) {
+                    archive.delete();
+                }
+
                 return true;
             }
